refactor(chat): extract orchestration into useAgentChat hook
Move all orchestration (load effect, pagination, polling, WS subscribe+
reconnect, send/retry/retryTask, nap, the orchestration refs + state, derived
memos) out of agent-chat-view.tsx into a new use-agent-chat.ts hook. The JSX
render stays in the component untouched, sourcing values from
`const chat = useAgentChat(props, external)`.

agent-chat-view.tsx 2797 -> 1175 lines. Zero behavior change: orchestration
moved verbatim (same dep arrays, same effect order, 6 ordering constraints
intact), JSX byte-identical. The one deliberate change is a useEffect->
useLayoutEffect on the slash-command ref bridge, so it populates before the
hook's passive load effect reads it (reproduces the original closure timing;
fixes a skill-draft-restore drop the ref bridge would otherwise introduce).

Interface: 5 pass-in setters + 4 useLatest reader-refs + draftMetaRestoredRef +
slashCommandRef/activeSkillRef bridge; scrollRef/composerRef created in hook and
returned; fileInputRef stays in useFileAttachments.

Verified: my review + code-reviewer + QA browser pass (all 5 ACs incl the
effect-ordering edge on both nav paths); web suite 1343/1343 at baseline.
